New ekoi (#10)
* Change int to long due to bigint type of size column of easy_files table. More info can be found here:
https://drivenbydata.atlassian.net/browse/EASY-981

* Bug fixed: popup screen is as dark as the background
More info can be found here:
https://drivenbydata.atlassian.net/browse/EASY-1050

+++ lib-deprecated/dans-lang/src/main/java/nl/knaw/dans/common/lang/service/exceptions/FileSizeException.java
@@ -3,8 +3,8 @@ package nl.knaw.dans.common.lang.service.exceptions;
 public class FileSizeException extends ServiceException {
     private static final long serialVersionUID = 8483469299362768976L;
 
-    protected int amount;
-    protected int limit;
+    protected long amount;
+    protected long limit;
 
     public FileSizeException(String message) {
         super(message);
@@ -21,17 +21,17 @@ public class FileSizeException extends ServiceException {
     /*
      * Note that sizes in MegaBytes
      */
-    public FileSizeException(int amount, int limit) {
+    public FileSizeException(long amount, long limit) {
         super("The file(s) exceeds the max size limit of " + limit + "megabytes");
         this.amount = amount;
         this.limit = limit;
     }
 
-    public int getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public int getLimit() {
+    public long getLimit() {
         return limit;
     }
